bug fix endpoint

+++ simple-online-shop/src/main/java/com/tujuhsembilan/simple_online_shop/model/Item.java
@@ -17,7 +17,7 @@ public class Item {
     @Id
     @Column(name = "item_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer itemId;
+    private Long itemId;
 
     @Column(name = "item_name")
     private String itemName;
